+++ service/rds/api.go
@@ -28287,7 +28287,7 @@ type DescribeExportTasksInput struct {
 	// Default: 100
 	//
 	// Constraints: Minimum 20, maximum 100.
-	MaxRecords *string `type:"string"`
+	MaxRecords *int64 `min:"20" type:"integer"`
 
 	// The Amazon Resource Name (ARN) of the snapshot exported to Amazon S3.
 	SourceArn *string `type:"string"`
@@ -28306,6 +28306,9 @@ func (s DescribeExportTasksInput) GoString() string {
 // Validate inspects the fields of the type to determine if they are valid.
 func (s *DescribeExportTasksInput) Validate() error {
 	invalidParams := request.ErrInvalidParams{Context: "DescribeExportTasksInput"}
+	if s.MaxRecords != nil && *s.MaxRecords < 20 {
+		invalidParams.Add(request.NewErrParamMinValue("MaxRecords", 20))
+	}
 	if s.Filters != nil {
 		for i, v := range s.Filters {
 			if v == nil {
@@ -28342,7 +28345,7 @@ func (s *DescribeExportTasksInput) SetMarker(v string) *DescribeExportTasksInput
 }
 
 // SetMaxRecords sets the MaxRecords field's value.
-func (s *DescribeExportTasksInput) SetMaxRecords(v string) *DescribeExportTasksInput {
+func (s *DescribeExportTasksInput) SetMaxRecords(v int64) *DescribeExportTasksInput {
 	s.MaxRecords = &v
 	return s
 }